Changed the way players can add/remove lighthouses on edges
Clicking in the polygon adds a lighthouse to the nearest line segment (in the middle for now). The lighthouse has the line segment  it is on saved as a variable.

If the closest line segment already contains a lighthouse, this lighthouse is removed (and no new lighthouse is added)

Dragging lighthouses was removed.