Bugfixes + new level
Fixed a bug in the intro level where two vertices were stacked on top of each other.
Fixed a bug where the resulting line segments were of the wrong orientation (180 degrees).
Fixed a bug where a line segments was created for a concave vertex that was colinear with two other points it can see and adjacent to one of those points.
Added another test level.