Merging Line Segment + Updated Controls Description
Updated the Controls Description to notify the user of the vertex, face, and edge toggle.

Implemented a MultiLineSegment, which is a line segment consisting of multiple other line segments. This class is used to merge together the generated line segments for the visibility reigons.
"Vertices do not lie on the same face"-exception is still not fixed though.