Vertex Epsilon distance changed + QoL changes
Massively increased the epsilon-values when comparing vertices (from 1e-4 to 1e-1). NB: This is still a relatively small number compared to the new upscaled polygon coordinates.

Rewrote the method that checks whether a vertex of a polygon is convex to now utilise caching (similar to several other polygon2D methods).

The game now throws an error if an invalid polygon is used for a level. An invalid polygon is a polygon of which two vertices are within vertex-EPS distance of each other.

Removed the general-position-randomisation from the Art Gallery LevelImporter.